Isolate SFX random handle behaviour
Background:
The SFX module has an accessor for the current sound handle, this is used to not only play the sound but also get info about it too. This accessor returns a random value for every call. This is fine when playing a new sound, but not for getting its info.

Changes:
* Added SFX_Current_GetSound3DHandle function. This returns the currently-assigned value, and never randomizes it. Meanwhile the random version will only be called when preparing to play a sound from within the module.
* Replaced instances of SFX_Random_GetSound3DHandle with Current where appropriate.
* Added checks for invalid sound3DHandles where they were missing.
* Changed existing and new checks for invalid sound3DHandles to use `value > 0`. Because 0 is treated as uninitialized, and -1 is considered invalid by the Sound3D module.