Fix #41: LWS sounds no longer trigger twice on frame 0
Background:
An issue popped up where suddenly sounds were played twice over each other. For random sounds this was incredibly noticeable (after fixing #37). This aims to solve that without creating any new bugs.

See the postmortem in #41 for detailed information on the issue. But the gist is that the check for hitting a frame to start playing a sound in LWS scenes was inclusive, meaning our lastTime could be equal, or our current time could. This became a problem for sounds with a trigger frame on 0. Because the initial values for both were 0 and 0, followed by 0, and <higher value> on the next call (meaning the trigger would be hit twice).

Checks were also done for sounds intended to trigger on the very last frame, and thankfully they always trigger when lastTime <= lastFrame and currTime >= 0.0. So it's safe to use the exclusive check for the upper bounds.

Changes:
* Added `Lws_KeyPassedExclusive` for used when trying to play sound triggers, this changes `<=` to `<` for instances of checking the upper bound of the time.
* Sound triggers now try to stop the current sound before playing a new one on the same loop cycle.
* loopUID now has its elements initialized to `-1` in the rare scenario that a sound stopped before played for the first time.

Refactors:
* Added parameter modifiers where appropriate to engine functions that were noticed while writing this fix.
* loopUID element type changed to signed integer, to properly support invalid sound play handles. (Every instance where loopUID was involved used signed integers, so there's no reason not to switch.)